Removed AddingMerchantService and AddingMerchantsServiceImplementation, added addMerchant method to MerchantService.
